feat!: add `systemd` launch mode for linux, and refactor argument passing for launch modes.
Now we use enums instead of booleans to pass launch mode for macOS and Linux, 
because I want to support more launch modes for Linux, such as crontab.

I add `systemd` launch mode because most of my use cases for rust projects to auto launch 
is on server instead of desktop environment, 
which means XDG method is not working.

Sorry for bringing a relatively big PR!